Fix cutils.isolate(): fix main loop, handling of opts.noHead, opts.noCmdline
The main loop was messed up when this function was updated. Put elements to keep
in keepElts, then loop through all elements to remove those not containing or
contained in an element to keep.

The opts.noHead and opts.noCmdline options were mishandled in a couple of ways.
(1) They were appended to the selector with a trailing comma, causing the
selector to be malformed.
(2) By being part of the selector, they became subject to opts.context,
opts.firstMatch, opts.match and opts.filter, any of which may remove them from
the elements to keep. Instead, they should be added directly to keepElts.